+++ src/minibuffer.rs
@@ -3,7 +3,7 @@
 
 use std::{
     io::{stdout, Write},
-    process::Output,
+    process::{Command, Output},
     str,
 };
 
@@ -26,6 +26,8 @@ pub struct MiniBuffer {
     current_height: usize,
     /// History of commands sent via `:`.
     git_command_history: Vec<String>,
+    /// History of commands sent via `!`.
+    command_history: Vec<String>,
 }
 
 pub enum MessageType {
@@ -66,32 +68,28 @@ impl MiniBuffer {
         }
     }
 
-    /// Call to enter Command mode. It will be exited automatically in the render call after the
-    /// command is sent.
-    pub fn git_command(&mut self, term_width: u16, term_height: u16) -> Result<()> {
-        // Clear the git output, if there is any.
-        print!(
-            "{}{}{}",
-            cursor::MoveTo(0, term_height.saturating_sub(self.current_height as u16)),
-            terminal::Clear(ClearType::FromCursorDown),
-            cursor::Show,
-        );
-
-        let mut command = String::with_capacity(term_width as usize - 5);
+    fn take_input<'a>(
+        prompt: &str,
+        term_width: u16,
+        term_height: u16,
+        history: &'a mut Vec<String>,
+    ) -> Option<&'a str> {
+        let mut input_buffer = String::new();
         let mut cursor = 0;
         let mut history_cursor = 0;
         loop {
             print!(
-                "{}{}\r\n{}:git {command}{}{}",
+                "{}{}\r\n{}{prompt}{command}{}{}",
                 cursor::MoveTo(0, term_height - 2),
                 "\u{2574}".repeat(term_width.into()),
                 terminal::Clear(ClearType::CurrentLine),
-                cursor::MoveToColumn(cursor + 5),
-                if command.len() == cursor.into() {
+                cursor::MoveToColumn(cursor + prompt.len() as u16),
+                if input_buffer.len() == cursor.into() {
                     SetCursorStyle::DefaultUserShape
                 } else {
                     SetCursorStyle::SteadyBar
-                }
+                },
+                command = input_buffer
             );
             drop(stdout().flush());
 
@@ -99,9 +97,7 @@ impl MiniBuffer {
             // same place as everywhere else (currently in `main.rs`). This would solve other minor
             // issues like the minibuffer not adjusting when the terminal font is increased. This
             // could probably be achieved by having the minibuffer be a `View`.
-            if let Event::Key(key_event) =
-                event::read().context("failed to read a terminal event")?
-            {
+            if let Event::Key(key_event) = event::read().expect("failed to read a terminal event") {
                 if key_event.kind == KeyEventKind::Release {
                     continue;
                 }
@@ -111,83 +107,118 @@ impl MiniBuffer {
                         cursor = cursor.saturating_sub(1);
                     }
                     (KeyCode::Right, _) | (KeyCode::Char('f'), KeyModifiers::CONTROL) => {
-                        if (cursor as usize) < command.len() {
+                        if (cursor as usize) < input_buffer.len() {
                             cursor += 1;
                         }
                     }
                     (KeyCode::Up, _) | (KeyCode::Char('p'), KeyModifiers::CONTROL) => {
-                        if history_cursor < self.git_command_history.len() {
+                        if history_cursor < history.len() {
                             history_cursor += 1;
-                            self.git_command_history
-                                [self.git_command_history.len() - history_cursor]
-                                .clone_into(&mut command);
-                            cursor = command.len() as u16;
+                            history[history.len() - history_cursor].clone_into(&mut input_buffer);
+                            cursor = input_buffer.len() as u16;
                         }
                     }
                     (KeyCode::Down, _) | (KeyCode::Char('n'), KeyModifiers::CONTROL) => {
                         history_cursor = history_cursor.saturating_sub(1);
                         if history_cursor == 0 {
-                            command.clear();
+                            input_buffer.clear();
                         } else {
-                            self.git_command_history
-                                [self.git_command_history.len() - history_cursor]
-                                .clone_into(&mut command);
+                            history[history.len() - history_cursor].clone_into(&mut input_buffer);
                         }
-                        cursor = command.len() as u16;
+                        cursor = input_buffer.len() as u16;
                     }
                     (KeyCode::Home, _) => cursor = 0,
-                    (KeyCode::End, _) => cursor = command.len() as u16,
+                    (KeyCode::End, _) => cursor = input_buffer.len() as u16,
                     (KeyCode::Char('b'), KeyModifiers::ALT) => {
                         while cursor > 0 {
                             cursor = cursor.saturating_sub(1);
-                            if word_boundary(&command, cursor as usize) {
+                            if word_boundary(&input_buffer, cursor as usize) {
                                 break;
                             }
                         }
                     }
                     (KeyCode::Char('f'), KeyModifiers::ALT) => {
-                        while cursor < command.len() as u16 {
+                        while cursor < input_buffer.len() as u16 {
                             cursor += 1;
-                            if word_boundary(&command, cursor as usize) {
+                            if word_boundary(&input_buffer, cursor as usize) {
                                 break;
                             }
                         }
                     }
                     (KeyCode::Char(c), _) => {
-                        command.insert(cursor.into(), c);
+                        input_buffer.insert(cursor.into(), c);
                         cursor += 1;
                     }
                     (KeyCode::Backspace, _) => {
                         if cursor > 0 {
                             cursor -= 1;
-                            command.remove(cursor.into());
+                            input_buffer.remove(cursor.into());
                         }
                     }
                     (KeyCode::Delete, _) => {
-                        if (cursor as usize) < command.len() || cursor == 0 && command.len() == 1 {
-                            command.remove(cursor.into());
-                        } else if !command.is_empty() {
-                            command.pop();
+                        if (cursor as usize) < input_buffer.len()
+                            || cursor == 0 && input_buffer.len() == 1
+                        {
+                            input_buffer.remove(cursor.into());
+                        } else if !input_buffer.is_empty() {
+                            input_buffer.pop();
                             cursor -= 1;
                         }
                     }
                     (KeyCode::Esc, _) => {
-                        print!("{}", cursor::Hide);
-                        return Ok(());
+                        input_buffer.clear();
+                        break;
                     }
                     _ => {}
                 }
             }
         }
+        (!input_buffer.is_empty()).then(|| {
+            history.push(input_buffer);
+            history.last().expect("we just pushed an elem").as_str()
+        })
+    }
+
+    /// Call to enter command mode. It will be exited automatically in the render call after the
+    /// command is sent.
+    pub fn command(&mut self, term_width: u16, term_height: u16, git_cmd: bool) -> Result<()> {
+        // Clear the git output, if there is any.
+        print!(
+            "{}{}{}",
+            cursor::MoveTo(0, term_height.saturating_sub(self.current_height as u16)),
+            terminal::Clear(ClearType::FromCursorDown),
+            cursor::Show,
+        );
+
+        let (prompt, history) = if git_cmd {
+            (":git ", &mut self.git_command_history)
+        } else {
+            ("!", &mut self.command_history)
+        };
+        let cmd = Self::take_input(prompt, term_width, term_height, history);
 
         crossterm::execute!(stdout(), cursor::MoveToColumn(0))?;
         terminal::disable_raw_mode().context("failed to disable raw mode")?;
-        self.push_command_output(&git_process(
-            &command.split_whitespace().collect::<Vec<_>>(),
-        )?);
+        if let Some(cmd) = cmd {
+            let cmd_output = if git_cmd {
+                let output = git_process(&cmd.split_whitespace().collect::<Vec<_>>());
+                Some(output)
+            } else {
+                let mut words = cmd.split_whitespace();
+                words.next().map(|cmd| {
+                    Command::new(cmd)
+                        .args(words.collect::<Vec<_>>())
+                        .output()
+                        .context("failed to run command")
+                })
+            };
+            match cmd_output {
+                Some(Ok(cmd_output)) => self.push_command_output(&cmd_output),
+                Some(Err(e)) => self.push(&format!("{e:?}"), MessageType::Error),
+                None => {}
+            }
+        }
         terminal::enable_raw_mode().context("failed to enable raw mode")?;
-
-        self.git_command_history.push(command);
 
         print!("{}", cursor::Hide);
         Ok(())